Added the Scene class
The scene class will be the dynamic processes in the game by switching between game states such as changing between the menu, the options and levels. It'll be cleaner to have a dedicated scene class with it's own subclasses for the different game states than to include it in program (as program is only in charge of running the game).
In the current version, the game only consists of switching between two background colors by pressing "w".
Please refer to updates.txt for a more detailed description of changes in v0.0.2.