Purple Square Movement and Platforms
The LevelScene class was updated to include the necessary platforms and walls for the current first level (as a test or to be used in the final game). Furthermore, there are input commands for that level where in the future, these are put into a main Level class that branches off into subclasses to avoid repetition (multiple levels).
Added SquareMe class representing the main playable character as a purple square. The SquareMe class should be complete with a constant left and right movement, jump action (space or w) and collision with walls and on top of platforms. Currently, there isn't collision for the bottom of platforms and is planned to be implemented in the next patch. 
"c" is used instead of "w" to switch between the prototype "menu" and "level" scene.