Main v.0.1.1
This should be version 11 (v.0.1.1) or not, if I've miscounted.
This update is mainly to start moving things into their own classes and eventually move them to their respective files.
I've moved music into it's own class and will be integrating it into the LevelScene class in the next update. File path also has been altered compared to this update, I'll be fixing that in the next update.