dsn_levels v.0.1.5
Major changes to the formatting of the platforms, walls, win and death zones to not be defined in individual temporary variables, but rather in the list directly. Also changes to the menu to include a placeholder title splash and 4 options that can be navigated with using a and d. A temporary Filler scene class was made to act as placeholders for the other 3 options.