dsn_class v.0.1.4
Added a memory class to record player statistics such as total deaths and jumps. The plan is to include the levels that the player had completed and allow the player to revisit them. 
Also, the SquareMe (square player) class has an additional jumps statistic to complement the memory class. This jumps variable is apart of the dsn_levels bug fix in holding jumps.