main v.0.1.5
Changed pause + quit (with q) to set self.run_scene to False instead of closing game to avoid crashing. Also altered main to complement these changes, allowing the game to be closed (not scene.run_scene) with a button press (in the levels) while the game is still running (self.running), seen in line 30.
Added function to the Memory class in main and in the Levels and level_id to Scene. Levels with id lesser than -1 are invalid levels (such as filler, or options/credits in the future). Level id's greater than -1 are actual levels with their own death/jump statistics.
Level ID is also used to track the player's level progress by adding the most recently played game to it's dictionary. Exact changes can be seen in the Memory class in dsn_class.
MenuScene now has a jump counter, but doesn't have a death counter (as you cannot die in the menu, yet).
Main was changed to have memory and record data. It will always record data in the menu as soon as the update function runs. In contrast, the data will only be recorded for levels once it has been beaten. Both are still inefficient processes as the menu is always recording and updating, and the levels have a burst in recording data (during the level transition).
There was an additional bug fix where the player would sink into the platform for each bounce it did.
This update is in addition to dsn_class and dsn_level v.0.1.6.