dsn_levels v.0.2.0
Fixed level text being in front of pause/win text.

Fixed gravity counter (amplification) not resetting when respawning. This issue was obvious when the player respawn was set in the air above platforms.

Player rendering has been temporarily put in the render_text functions. The player will still be rendered behind text, but will be in front of level elements. In the future, this function could be renamed to "render_over", as in rendering over the elements. This is an important change for rendering the player above the respawn block.

Player respawn has also been added, showing up as lime green. Currently, the respawn for the player colliding with these elements is set to be in the middle of them. Adding respawns need to be done in a "self.respawn_zones" list within each levels "level_render" class function. The new respawn feature could be seen in the "Mind the Gap" level, or what it's currently set to.