dsn_levels v.0.2.3
Added a "PROOF OF CONCEPT" level select screen with it's animation. It only has one speed for choosing levels.

!!!!
IMPORTANT: 
There is no function to the level selector yet, as there's no way to keep track of all of the levels (like a list of all the levels or a text file containing all of the levels). Therefore, this function will be added when the levels text file gets updated.

To start the game as usual, just press jump ("space", "up" or "w"), which will start you at level 1.
!!!!

Please let me know if you want to test a set of levels in the future and put them as a global list. I will add this option to the level selector