dsn_class v.0.1.9
Added color constants back to dsn_class. Also created DSNElement to hold information for drawing a pygame rect, including the color and dimensions. Both are necessary in loading information from a text file.
The text file is being loaded in the Memory class with the "load_levels" method/function.
The output is seen with "print_levels", already put into "load_levels".

Finally, the local element_lookup and color_lookup variables are defined but not used. This leaves the option for converting level element identifiers from strings into integers if loading levels in dsn_levels isn't compatible or as readable with strings.