dsn_levels v.0.2.9
Added a statistics page that will display the death counter and jump counter for that level if completed successfully. In the future, it will display total deaths, total jumps and other planned statistics. It also needs more decoration to be visually appealing, as currently, it only has text with no decoration.

Other small changes to complement this update is that LevelScene now takes memory as a parameter, rather than level data/elements contained in memory.