main v.0.2.1
Mini update:

1. Moved statistics updating (update_mem) and level changing (self.level_id) to dsn_levels from main. 

This is related to issue #204 as there would be 2 blocks of code that occur independently when the level was completed. There was a small chance that the level transition in dsn_levels would occur before the statistics/level increment (switch levels), causing the level would repeat itself. 
This issue shouldn't occur anymore (or has a drastically reduced chance to) because all things related to level changing was moved to 1 block of code.

-----=-----

2. Removed music value in preparation for having music play in a wider scope, or in the main loop in main. 

This is related to issue #185 where the music would be reinitialized every time the level was completed. When music is moved to main (under the Program class or above in scope), music would be continuous as there wouldn't be anything interrupting/restarting it.
@shoc71 We should consider what music would play during the levels (theme) and when the music would play (is it the same every time, or is it random for that level)