dsn_levels v.0.3.7 fix 1
Fixed a spelling mistake in a comment and fully fixed jumping. I forgot to change the timing for held jumps from 200 to 150 milliseconds. 
The bug was related to jump boost where on respawn, the player square would still be jumping. The fix was that the jump boost wasn't reset to it's default value on death, causing the jump to continue when the player respawned.